client/requirements.txt
PyYAML==5.3.1
pytest==5.3.2
pygame==2.0.0.dev6
